[radix cache split] Address review: msgspec actions, CI-active cap tests, exception-safe drain
- Convert cache_action.py containers from NamedTuple/dataclass to msgspec.Struct(frozen=True)
- Wrap the deferred mamba path-cap walk in try/finally so collected frees drain on a mid-walk failure; add a fault-injection regression test
- Move the write-through/path-cap ordering regressions into test_mamba_path_state_cap.py and register that module for cuda CI (mamba pool fixtures need CUDA; the class skips on the cpu suite)
- Fix the lock benchmark's NodeId-vs-node root comparison
- Update the components README to the controller/tree-core split; add the tree-core file-split TODO
- isort the touched files
