+++ test/registered/unit/mem_cache/test_mamba_path_state_cap.py
@@ -1,15 +1,19 @@
 """CPU-only unit tests for the per-path Mamba checkpoint cap."""
 
-from sglang.test.ci.ci_register import register_cpu_ci
+from sglang.test.ci.ci_register import register_cpu_ci, register_cuda_ci
 
 register_cpu_ci(est_time=2, suite="base-a-test-cpu")
+register_cuda_ci(est_time=5, stage="base-b", runner_config="1-gpu-small")
 
 import argparse
-from collections import defaultdict
 import unittest
+from collections import defaultdict
+from unittest import mock
 
 import torch
+from test_unified_radix_cache_unittest import CacheConfig, UnifiedRadixCacheSuite
 
+from sglang.srt.mem_cache.unified_cache.cache_action import MambaEvictExcessPathStates
 from sglang.srt.mem_cache.unified_cache.unified_tree_core import UnifiedTreeCore
 from sglang.srt.mem_cache.unified_cache_components.mamba_component import (
     MambaComponent,
@@ -19,6 +23,7 @@ from sglang.srt.mem_cache.unified_cache_components.tree_component import (
 )
 from sglang.srt.mem_cache.unified_radix_cache import UnifiedLRUList, UnifiedTreeNode
 from sglang.srt.server_args import ServerArgs
+from sglang.test.test_utils import CustomTestCase
 
 
 class _FakeTreeCore:
@@ -175,6 +180,137 @@ class TestMambaPathStateCap(unittest.TestCase):
                 node.component_data[ComponentType.MAMBA].value is not None
                 for node in nodes
             )
+        )
+
+
+@unittest.skipUnless(torch.cuda.is_available(), "mamba pool fixtures need CUDA")
+class TestMambaPathCapWriteThroughOrdering(CustomTestCase):
+    """CI-active write-through/path-cap ordering regressions (the unified radix
+    cache unittest module is temporarily gated off on trunk)."""
+
+    cfg = CacheConfig(components=(ComponentType.FULL, ComponentType.MAMBA))
+    _rid = 0
+    # Borrow the fixture helpers without inheriting the full gated suite.
+    _make_req = UnifiedRadixCacheSuite._make_req
+    _alloc = UnifiedRadixCacheSuite._alloc
+    _insert = UnifiedRadixCacheSuite._insert
+    _init_hicache = UnifiedRadixCacheSuite._init_hicache
+    _build_hicache_fixture = UnifiedRadixCacheSuite._build_hicache_fixture
+
+    def test_write_through_backup_survives_mamba_path_cap(self):
+        cache, allocator, req_to_token_pool = self._build_hicache_fixture()
+        cache.write_through_threshold = 2
+        cache.components[ComponentType.MAMBA].mamba_max_states_per_path = 1
+
+        # The first insert creates the ancestor with a mamba state (hit_count 1).
+        self._insert(cache, allocator, req_to_token_pool, [1, 2])
+        ancestor = next(iter(cache.root_node.children.values()))
+        self.assertIsNotNone(ancestor.component_data[ComponentType.MAMBA].value)
+
+        # The extending insert crosses the ancestor's write-through threshold in
+        # the same walk whose commit runs the path-cap eviction; the cap must
+        # leave the pending-backup node's device state for the deferred BackupKV.
+        self._insert(cache, allocator, req_to_token_pool, [1, 2, 3, 4])
+        cache.writing_check(write_back=True)
+
+        self.assertTrue(ancestor.backuped)
+        self.assertIsNotNone(ancestor.component_data[ComponentType.MAMBA].host_value)
+
+    def test_write_through_backup_chain_survives_mamba_path_cap(self):
+        """A failed backup leaves an unbacked ancestor inside a later deferred
+        backup chain; the cap walk must spare the whole chain, not just its tip."""
+        cache, allocator, req_to_token_pool = self._build_hicache_fixture()
+        cache.write_through_threshold = 3
+        mamba_comp = cache.components[ComponentType.MAMBA]
+
+        self._insert(cache, allocator, req_to_token_pool, [1, 2])
+        ancestor = next(iter(cache.root_node.children.values()))
+        self._insert(cache, allocator, req_to_token_pool, [1, 2, 3, 4])
+        middle = next(iter(ancestor.children.values()))
+
+        # The ancestor crosses the threshold here; a host-exhaustion failure
+        # leaves it unbacked with hit_count past the bar and its state intact.
+        with mock.patch.object(cache, "_execute_kv_backup", return_value=None):
+            self._insert(cache, allocator, req_to_token_pool, [1, 2, 3, 4, 5, 6])
+        self.assertFalse(ancestor.backuped)
+        self.assertIsNotNone(ancestor.component_data[ComponentType.MAMBA].value)
+
+        # The middle node crosses next, so the deferred chain is
+        # [ancestor, middle]; the extending insert adopts a new leaf state,
+        # firing the now-enabled cap walk before the chain executes — it must
+        # not evict either chain node's device state.
+        mamba_comp.mamba_max_states_per_path = 1
+        self._insert(cache, allocator, req_to_token_pool, [1, 2, 3, 4, 5, 6, 7, 8])
+        cache.writing_check(write_back=True)
+
+        self.assertTrue(ancestor.backuped)
+        self.assertIsNotNone(ancestor.component_data[ComponentType.MAMBA].host_value)
+        self.assertTrue(middle.backuped)
+        self.assertIsNotNone(middle.component_data[ComponentType.MAMBA].host_value)
+
+    def test_backup_retry_after_mamba_cap_skips_tombstoned_state(self):
+        """A backup that fails before the cap and retries via the leaf action
+        rebuilds its spec post-cap: KV backs up, the tombstoned mamba arm stays gone."""
+        cache, allocator, req_to_token_pool = self._build_hicache_fixture()
+        cache.write_through_threshold = 1
+        mamba_comp = cache.components[ComponentType.MAMBA]
+
+        # A failed write-through leaves the ancestor unbacked with device state.
+        with mock.patch.object(cache, "_execute_kv_backup", return_value=None):
+            self._insert(cache, allocator, req_to_token_pool, [1, 2])
+        ancestor = next(iter(cache.root_node.children.values()))
+        self.assertFalse(ancestor.backuped)
+        self.assertIsNotNone(ancestor.component_data[ComponentType.MAMBA].value)
+
+        # The walk backup fails again, the cap tombstones the unlocked
+        # ancestor's mamba state, then the leaf-action retry succeeds.
+        mamba_comp.mamba_max_states_per_path = 1
+        real_backup = cache._execute_kv_backup
+        attempts = []
+
+        def fail_once(*args, **kwargs):
+            attempts.append(args)
+            if len(attempts) == 1:
+                return None
+            return real_backup(*args, **kwargs)
+
+        with mock.patch.object(cache, "_execute_kv_backup", side_effect=fail_once):
+            self._insert(cache, allocator, req_to_token_pool, [1, 2, 3, 4])
+        leaf = next(iter(ancestor.children.values()))
+        cache.writing_check(write_back=True)
+
+        # Post-cap spec rebuild: no resurrection of the tombstoned mamba state.
+        self.assertTrue(ancestor.backuped)
+        ancestor_cd = ancestor.component_data[ComponentType.MAMBA]
+        self.assertIsNone(ancestor_cd.value)
+        self.assertIsNone(ancestor_cd.host_value)
+        self.assertTrue(leaf.backuped)
+        self.assertIsNotNone(leaf.component_data[ComponentType.MAMBA].host_value)
+        cache.sanity_check()
+
+    def test_cap_walk_failure_still_drains_collected_frees(self):
+        """A cap walk that raises mid-eviction must still free the tombstoned
+        slots it already collected (the pre-split inline frees could not leak)."""
+        cache, allocator, req_to_token_pool = self._build_hicache_fixture()
+        mamba_comp = cache.components[ComponentType.MAMBA]
+
+        self._insert(cache, allocator, req_to_token_pool, [1, 2])
+        ancestor = next(iter(cache.root_node.children.values()))
+        self._insert(cache, allocator, req_to_token_pool, [1, 2, 3, 4])
+        leaf = next(iter(ancestor.children.values()))
+        self.assertIsNotNone(ancestor.component_data[ComponentType.MAMBA].value)
+
+        mamba_comp.mamba_max_states_per_path = 1
+        available = req_to_token_pool.mamba_allocator.available_size()
+        with mock.patch.object(
+            cache.tree_core, "_cascade_evict", side_effect=RuntimeError("boom")
+        ):
+            with self.assertRaises(RuntimeError):
+                mamba_comp.apply_component_action(MambaEvictExcessPathStates(leaf.id))
+
+        self.assertIsNone(ancestor.component_data[ComponentType.MAMBA].value)
+        self.assertEqual(
+            req_to_token_pool.mamba_allocator.available_size(), available + 1
         )
